✅create `test_editgridschema.py` and check tests are working
